* x11vnc-updates:
  Non-blocking sockets for Windows.
  Cleanup: remove CORBA stuff.
  update to x11vnc 0.9.12
  x11/x11vnc.c: include Xproto.h as early as possible

+++ ica/x11/x11vnc/inet.c
@@ -347,7 +347,7 @@ char *ident_username(rfbClientPtr client) {
 			if (ret > 0) {
 				int i;
 				char *q, *p;
-				for (i=0; i < sizeof(msg); i++) {
+				for (i=0; i < (int) sizeof(msg); i++) {
 					msg[i] = '\0';
 				}
 				usleep(250*1000);
